feat: add climbs table
Add climbs table. The climbs are made up of just a name for now. These
names can be null, indicating they are unnamed. If they have a name, the
name should not be empty. The meaning of this would be ambiguous, as
no-name is represented by a null value.

Schema tests for this table have been added to t/test-climbs.pg and can
be ran with pgTAP's `pg_prove` program.