ci: migrate postgres service setup to docker-compose test workflow
As this project will be pulling in several extensions as time goes on,
this patch pre-emptively replaces the GitHub Action service container
with one created through docker-compose. This will allow us to install
extensions in the image, which we otherwise wouldn't be able to do
cleanly with the GitHub hosted runner service containers.

This also fixes a bit of misconfiguration where we could possibly have
been installing an incompatible version of pgTAP to postgres instance.
With this patch we use the inherent PG_MAJOR variable to install the
correct version of pgTAP. The pgTAP installed on the host runner is only
to access `pg_prove`, so that is okay to stay.