feat: add climb-verm-grades table
This patch adds a table that is used to assign vermin grades to rows in
the climbs table. This is a separate table to more keep the grade
attributes separate from the climbs as they could change more frequently
or many types of grades may be associated with a single climb.