test: verify function and trigger deployments
This uses the built-in function
[`pg_get_functiondef`](https://www.postgresql.org/docs/current/functions-info.html)
to verify function existence. It throws if the function doesn't exist.
Triggers are a bit tricker since they are made up a many things
including a particular table and function. For this a PL/pgSQL function
is used where a query is made against the [`pg_trigger`
table](https://www.postgresql.org/docs/current/catalog-pg-trigger.html)
and asserted for existence.