feat: add ascents table
This is a table that tracks an ascent for a particular climb. One thing
worth noting on this table is the `window` column is a `daterange`. This
is intended to be used as some sort of ambiguous time-point.. like if an
ascent happened in in January, you'd add a value like `'[2024-01-01,
2024-02-01)'`. If an ascent happened on an exact day then you'd add
something like `'[2024-01-01, 2024-01-02)'`. Some could interpret this
as an ascent occurred over several days, but this is not how this column
is intended to be used. In this scenario the date of the ascent finished
should be used and the rest of the information should be logged
elsewhere.