feat: add s3-image-sources table
This table represents s3 sources for a given image. It'd be expected
that if you have multiple sources for an image they would give you the
same image, or at least conceptually the same image (like maybe
different sizes for a picture).

Records for this table should be inserted through events of an s3 store.
This way the database is always represents the s3 store. It follows that
instead of cascading on delete, records in this table restrict the
deletion of records in the images table. This way you won't "orphan" an
s3 object by deleting records in the database.