refactor!: relax mutual exclusivity checks
When a closure from a to b exists, one may want to attempt to insert
another closure from a to b with the same primary key and perform an
update when the primary-key conflict occurs. This currently isn't
possible because the trigger function checks all possible parents
instead so the trigger function will raise before the primary-key
constraint will.

This patch splits the responsibility of the trigger function into
multiple trigger functions that only raise if one table contains a
matching row, effectively relaxing the exclusivity pre-check condition.

This does change expected, likely edge-case, behavior slightly.