feat: add topo image features table
Rows in this table represent images that can be put on a topo. The
images can be cropped into any place through the boxes of the
`source_crop` and `dest_crop` columns. The dimensions of these boxes are
determined by the callers, though some sensible options could be
normalized coordinates or source being in the frame of the implicit
image size and the destination being in the frame of the topo size.