feat(formations): add "parent" foreign keys
A formation can be long to a region, a crag, a sector, or none,
exclusively. This patch adds a column for each of the entities and a
named constraint check to enforce this "at most one parent" idea. This
check is necessary as it prevents nonsensical data from entering the
table.

Note that the test does not check for a named constraint as that is
(currently) a limitation of pgTAP.